projects/auto-qm-form/pyproject.toml
[project]
name = "auto-qm-form"
version = "0.1.0"
description = "Auto QM Form"
requires-python = ">=3.11"
dependencies = [
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
  "pre-commit",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.coverage.run]
branch = true
source = ["src"]